warning

+++ etqw/AtmosphereRenderable.cpp
@@ -271,7 +271,7 @@ void sdAtmosphereRenderable::UpdateCelestialBody( parms_t& parms ) {
 	skyLightGlowSprite.callback = glowSpriteCB;
 #pragma warning( push )
 #pragma warning( disable: 4312 )
-	skyLightGlowSprite.callbackData = (void*)Uid;
+	skyLightGlowSprite.callbackData = (void*)(uintptr_t/*//k64 */)Uid;
 #pragma warning( pop )
 
 	skyLightGlowSprite.customShader = parms.atmosphere->GetSunFlareMaterial();
@@ -454,7 +454,7 @@ bool sdAtmosphereRenderable::glowSpriteCB( renderEntity_t *re, const renderView_
 	if ( v ) {
 #pragma warning( push )
 #pragma warning( disable: 4311 )
-		sdAtmosphereRenderable *atmos = static_cast<sdAtmosphereRenderable *>(renderSystem->PtrForUID( (intptr_t/*//k 64 int*/)re->callbackData ));
+		sdAtmosphereRenderable *atmos = static_cast<sdAtmosphereRenderable *>(renderSystem->PtrForUID( (uintptr_t/*//k 64 int*/)re->callbackData ));
 #pragma warning( pop )
 		if ( atmos != NULL ) {
 			return atmos->_glowSpriteCB( re, v, lastModifiedGameTime );
